+++ src/envConfig.ts
@@ -425,6 +425,11 @@ export const asEnvConfig = asObject({
       quiknodeApiKey: asOptional(asString, '')
     }).withRest
   ),
+  STEALTHEX_INIT: asCorePluginInit(
+    asObject({
+      apiKey: asOptional(asString, '')
+    }).withRest
+  ),
   SWAPTER_INIT: asCorePluginInit(
     asObject({
       apiKey: asOptional(asString, '')